Réalisation de la class View(Camera) et debut de la class Game/Modification de la class Vector

+++ src/SuperOpenGL/Vector.cpp
@@ -6,18 +6,18 @@ namespace SuperOpenGL {
 
 Vector::Vector()
 {
-  // modif RR, 05/12/2006, passage 0.0 -> 0.0f
-  x = 0.0;
-  y = 0.0;
-  z = 0.0;
+        // modif RR, 05/12/2006, passage 0.0 -> 0.0f
+        x = 0.0;
+        y = 0.0;
+        z = 0.0;
 }
 
-// modif RR 02/12/2006, parametre const long double -> const long double&
-Vector::Vector(const long double& X, const long double& Y, const long double& Z)
+// modif RR 02/12/2006, parametre const float -> const float&
+Vector::Vector(const float& X, const float& Y, const float& Z)
 {
-  x = X;
-  y = Y;
-  z = Z;
+        x = X;
+        y = Y;
+        z = Z;
 }
 
 /*!
@@ -26,9 +26,9 @@ Vector::Vector(const long double& X, const long double& Y, const long double& Z)
  */
 Vector::Vector(const Vector& _v)
 {
-  x = _v.x;
-  y = _v.y;
-  z = _v.z;
+        x = _v.x;
+        y = _v.y;
+        z = _v.z;
 }
 
 /*!
@@ -37,7 +37,7 @@ Vector::Vector(const Vector& _v)
  */
 Vector::~Vector()
 {
-  // rien
+        // rien
 }
 
 /*!
@@ -46,9 +46,9 @@ Vector::~Vector()
  */
 Vector::Vector(const Vertex& _P)
 {
-  x = _P.x;
-  y = _P.y;
-  z = _P.z;
+        x = _P.x;
+        y = _P.y;
+        z = _P.z;
 }
 
 /*!
@@ -57,46 +57,46 @@ Vector::Vector(const Vertex& _P)
  */
 Vector::Vector(const Vertex& _A, const Vertex& _B)
 {
-  x = _B.x - _A.x;
-  y = _B.y - _A.y;
-  z = _B.z - _A.z;
+        x = _B.x - _A.x;
+        y = _B.y - _A.y;
+        z = _B.z - _A.z;
 }
 
 bool
 Vector::operator==(const Vector& op) const
 {
-  return (x == op.x && y == op.y && z == op.z);
+        return (x == op.x && y == op.y && z == op.z);
 }
 
 Vector&
 Vector::operator=(const Vector& op)
 {
-  x = op.x;
-  y = op.y;
-  z = op.z;
-  return *this;
+        x = op.x;
+        y = op.y;
+        z = op.z;
+        return *this;
 }
 
 Vector&
 Vector::operator=(const Vertex& op)
 {
-  x = op.x;
-  y = op.y;
-  z = op.z;
+        x = op.x;
+        y = op.y;
+        z = op.z;
 
-  return *this;
+        return *this;
 }
 
 Vector
 Vector::operator+() const
 {
-  return (*this);
+        return (*this);
 }
 
 Vector
 Vector::operator+(const Vector& op) const
 {
-  return (Vector(x + op.x, y + op.y, z + op.z));
+        return (Vector(x + op.x, y + op.y, z + op.z));
 }
 
 /*! RR 02Juillet2008
@@ -104,120 +104,120 @@ Vector::operator+(const Vector& op) const
 Vector
 Vector::operator+(const Vertex& _p) const
 {
-  return (Vector(x + _p.x, y + _p.y, z + _p.z));
+        return (Vector(x + _p.x, y + _p.y, z + _p.z));
 }
 
 Vector&
 Vector::operator+=(const Vector& op)
 {
-  x += op.x;
-  y += op.y;
-  z += op.z;
-  return *this;
+        x += op.x;
+        y += op.y;
+        z += op.z;
+        return *this;
 }
 
 Vector
 Vector::operator-() const
 {
-  return (Vector(-x, -y, -z));
+        return (Vector(-x, -y, -z));
 }
 
 Vector
 Vector::operator-(const Vector& op) const
 {
-  return (Vector(x - op.x, y - op.y, z - op.z));
+        return (Vector(x - op.x, y - op.y, z - op.z));
 }
 
 Vector&
 Vector::operator-=(const Vector& op)
 {
-  x -= op.x;
-  y -= op.y;
-  z -= op.z;
-  return *this;
+        x -= op.x;
+        y -= op.y;
+        z -= op.z;
+        return *this;
 }
 
 Vector&
-Vector::operator*=(const long double op)
+Vector::operator*=(const float op)
 {
-  x *= op;
-  y *= op;
-  z *= op;
-  return *this;
+        x *= op;
+        y *= op;
+        z *= op;
+        return *this;
 }
 
-Vector Vector::operator*(const long double op) const
+Vector Vector::operator*(const float op) const
 {
-  return (Vector(x * op, y * op, z * op));
+        return (Vector(x * op, y * op, z * op));
 }
 
 Vector&
-Vector::operator/=(const long double op)
+Vector::operator/=(const float op)
 {
-  x /= op;
-  y /= op;
-  z /= op;
-  return *this;
+        x /= op;
+        y /= op;
+        z /= op;
+        return *this;
 }
 
 Vector
-Vector::operator/(const long double op) const
+Vector::operator/(const float op) const
 {
-  return (Vector(x / op, y / op, z / op));
+        return (Vector(x / op, y / op, z / op));
 }
 
-long double Vector::operator*(const Vector& op) const
+float Vector::operator*(const Vector& op) const
 {
-  return (x * op.x + y * op.y + z * op.z);
+        return (x * op.x + y * op.y + z * op.z);
 }
 
 Vector&
 Vector::operator^=(const Vector& op)
 {
-  Vector temp(y * op.z - z * op.y, z * op.x - x * op.z, x * op.y - y * op.x);
-  x = temp.x;
-  y = temp.y;
-  z = temp.z;
-  return *this;
+        Vector temp(y * op.z - z * op.y, z * op.x - x * op.z, x * op.y - y * op.x);
+        x = temp.x;
+        y = temp.y;
+        z = temp.z;
+        return *this;
 }
 
 Vector
 Vector::operator^(const Vector& op) const
 {
-  return (
-    Vector(y * op.z - z * op.y, z * op.x - x * op.z, x * op.y - y * op.x));
+        return (
+                       Vector(y * op.z - z * op.y, z * op.x - x * op.z, x * op.y - y * op.x));
 }
 
 Vector
-Vector::normalize(const long double lg)
+Vector::normalize(const float lg)
 {
-  long double lgtemp;
-  if (x != 0 || y != 0 || z != 0) {
-    lgtemp = sqrt(x * x + y * y + z * z);
-    lgtemp = lg / lgtemp;
-    x *= lgtemp;
-    y *= lgtemp;
-    z *= lgtemp;
-  }
-  return (*this);
+        float lgtemp;
+        if (x != 0 || y != 0 || z != 0) {
+                lgtemp = sqrt(x * x + y * y + z * z);
+                lgtemp = lg / lgtemp;
+                x *= lgtemp;
+                y *= lgtemp;
+                z *= lgtemp;
+        }
+        return (*this);
 }
 
-long double
+float
 Vector::length() const
 {
-  return (sqrt(x * x + y * y + z * z));
+        return (sqrt(x * x + y * y + z * z));
 }
 
 std::ostream&
 operator<<(std::ostream& os, const Vector& v)
 {
-  return os << v.x << ' ' << v.y << ' ' << v.z;
+        return os << v.x << ' ' << v.y << ' ' << v.z;
 }
 
 std::istream&
 operator>>(std::istream& is, Vector& v)
 {
-  return is >> v.x >> v.y >> v.z;
+        return is >> v.x >> v.y >> v.z;
 }
 
 }
